Split API out of dragonfly-script.py
